requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
cssselect2==0.5.0
Django==4.0.3
django-appconf==1.0.5
django-colorfield==0.6.3
django-compressor==3.1
django-fixture-magic==0.1.5
django-gunicorn==0.1.1
django-libsass==0.9
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
libsass==0.21.0
lxml==4.8.0
Markdown==3.3.6
Pillow==9.0.1
PyPDF3==1.0.6
python-bidi==0.4.2
PyYAML==6.0
rcssmin==1.1.0
reportlab==3.6.8
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
tqdm==4.63.0
webencodings==0.5.1
xhtml2pdf==0.2.6
